requirements.txt
pandas==1.4.3
scikit-learn==0.23.2
streamlit==1.11.1
pickle4==0.0.1
seaborn==0.7.1
plotly
pandas_datareader
keras.models
statsmodels
pip install TA-Lib
